Refactor into separate base class and S3KinesisFirehose implementation
