+++ lib/shortcuts/index.js
@@ -10,6 +10,6 @@ module.exports = {
   CrossAccountRole: require('./cross-account-role'),
   ServiceRole: require('./service-role'),
   Queue: require('./queue'),
-  KinesisFirehose: require('./kinesis-firehose'),
+  S3KinesisFirehose: require('./s3-kinesis-firehose'),
   hookshot: require('./hookshot')
 };
